+++ source/extensions/omni.flux.property_widget_builder.model.usd/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.13.0]
+### Added
+- Added identifier for UI elements
+
 ## [2.12.0]
 ### Changed
 - Use centralized LayerTree widget